pyproject.toml
[tool.poetry]
name = "omorpg"
version = "1.0.0"
description = "A discord bot that is OMORI Themed"
authors = ["LetsChill <undefined@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord = "^2.2.3"
python-dotenv = "^1.0.0"
motor = "^3.1.2"
jishaku = {git = "https://github.com/Gorialis/jishaku"}


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
